requirements.txt
streamlit==1.28.1
folium==0.14.0
pandas==2.1.3
requests==2.31.0
pyyaml>=6.0
geopandas==0.14.1
shapely==2.0.2
plotly==5.17.0
streamlit-folium==0.13.0
numpy==1.24.3

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Development dependencies
pre-commit==3.5.0 